add fully functional UI

+++ Pierres_Bakery/Models/BakedGood.cs
@@ -9,7 +9,7 @@ namespace Pierres_Bakery.Models
 
         public static float CalculatePrice(int itemNum, int price, int discountedItem, float discount)
         {
-            return ((float)itemNum * price) - ((float)itemNum / discountedItem * (price * discount));
+            return (itemNum * price) - (itemNum / discountedItem * (price * discount));
         }
     }
 }
